add construct
Tentative addition of a construct function by which to pass some default
values in parameters to the class itself.

Currently the ability to restrict min/max length of password is being
worked in. However, I can see how this could lead to problems, so it may
not stay for long...